+++ src/components/Dropdown/FilterDropdown/index.ts
@@ -1,0 +1,1 @@
+export { FilterDropdown } from './FilterDropdown'
